+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.9.0  
+##### 2020-04-30  
+  
+Add middleware and stream.  
+Everything is documented.  
+  
+  
 # 0.1.0  
 ##### 2020-04-30